refactor: move some shortcut files
