+++ rest_framework_supertest/shortcuts/__init__.py
@@ -129,6 +129,7 @@ from .emojis import emoji, unique_emoji
 from .files import (
     file_extension,
     file_name,
+    image,
     mime_type,
     unique_file_extension,
     unique_file_name,
@@ -142,7 +143,6 @@ from .geographicals import (
     unique_latitude,
     unique_longitude,
 )
-from .images import image
 from .internet import (
     domain_name,
     domain_word,
@@ -300,12 +300,10 @@ __all__ = [
     'emoji', 'unique_emoji',
     # files
     'file_extension', 'file_name', 'mime_type', 'unique_file_extension',
-    'unique_file_name', 'unique_mime_type',
+    'unique_file_name', 'unique_mime_type', "image",
     # geographical
     'coordinate', 'latitude', 'longitude', 'unique_coordinate', 'unique_latitude',
     'unique_longitude',
-    # images
-    "image",
     # internet
     'domain_name', 'unique_domain_name', 'domain_word', 'unique_domain_word', 'tld',
     'ipv4', 'unique_ipv4', 'ipv4_network_class', 'ipv4_private', 'unique_ipv4_private',
