+++ tests/models.py
@@ -9,3 +9,12 @@ class ImageModel(models.Model):
     def __str__(self):
         """Return a representation string of the model."""
         return str(self.field)
+
+class FileModel(models.Model):
+    """Define a Model to test setting file field."""
+
+    field = models.FileField()
+
+    def __str__(self):
+        """Return a representation string of the model."""
+        return str(self.field)
